backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["DonOutcast <sham1996@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg-binary = "^3.2.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
